+++ c2dataviewer/view/image_display.py
@@ -15,7 +15,7 @@ PVA object viewer utilities for image display
 from collections import namedtuple
 import queue
 import logging
-
+import math
 import numpy as np
 from pyqtgraph import QtCore
 from pyqtgraph.Qt import QtWidgets
@@ -76,6 +76,43 @@ class ImageCompressionUtility:
         nBytesWritten = blosc.decompress_ptr(bytearray(inputArray), outputArray.__array_interface__['data'][0])
         return outputArray
 
+class MouseDialog:
+    def __init__(self, imageWidget):
+        # mouse dialog flags        
+        self.mouse_dialog_enabled = False
+        self.mouse_dialog_launched = False
+        self.mouse_dialog_hide = False
+        
+        self.max_textbox_width = 0
+
+        # pixel coordinates
+        self.pix_x = 0
+        self.pix_y = 0
+
+        # textbox coordinates inside the widget screen
+        self.x_coor = 0
+        self.y_coor = 0
+
+        # intensity coordinates
+        self.int_x = 0
+        self.int_y = 0
+
+        # initialize mouse dialog widgets
+        self.layout = QtWidgets.QVBoxLayout(imageWidget)
+        self.textbox = QtWidgets.QTextEdit(imageWidget)
+        self.textbox.setCursorWidth(0)
+        self.textbox.setFixedSize(0,0)
+
+    def enable_mouse_dialog(self):
+        self.mouse_dialog_enabled = True
+        self.textbox.setHidden(False)
+
+    def disable_mouse_dialog(self):
+        self.mouse_dialog_enabled = False
+        self.mouse_dialog_launched = False
+        self.textbox.setHidden(True)
+    
+
 Image = namedtuple("Image", ['id', 'new', 'image', 'black', 'white'])
 class ImagePlotWidget(RawImageWidget):
 
@@ -86,6 +123,7 @@ class ImagePlotWidget(RawImageWidget):
 
     _set_image_signal = QtCore.pyqtSignal()
     connection_changed_signal = QtCore.Signal(str, str)
+    right_button_clicked_signal = QtCore.pyqtSignal(QtCore.QPoint)
     
     def __init__(self, parent=None, **kargs):
         RawImageWidget.__init__(self, parent=parent, scaled=True)
@@ -139,6 +177,8 @@ class ImagePlotWidget(RawImageWidget):
             'width': 0,
             'height': 0,
         }
+        self.right_button_pressed = False
+        self.is_image_panning = False
 
         # ROI
         self.roi_origin = None
@@ -199,6 +239,8 @@ class ImagePlotWidget(RawImageWidget):
         # Queue used to transfer the images from the process to the display thread
         self.draw_queue = queue.Queue(ImagePlotWidget.DEFAULT_DISPLAY_QUEUE_SIZE)
 
+        # mouse dialog box
+        self.mouse_dialog = MouseDialog(self)
 
     def resizeEvent(self, event):
         """
@@ -219,18 +261,27 @@ class ImagePlotWidget(RawImageWidget):
         """
         # Get location of the click
         click_position = event.pos()
-        x_position = click_position.x()
-        y_position = click_position.y()
+        self.click_x = click_position.x()
+        self.click_y = click_position.y()
+        
+        # Only triggered is left mouse button clicked 
+        if event.button() == QtCore.Qt.LeftButton:
+            # Return if there is no image data
+            if not self.x or not self.y:
+                return
+            
+            # Check if the press happened on the image
+            img_width, img_height, _ = self.calc_img_size_on_screen()
+            if (self.click_x > img_width or self.click_y > img_height):
+                return
 
-        # Check if the press happened on the image
-        img_width, img_height = self.calc_img_size_on_screen()
-        if (x_position > img_width or y_position > img_height):
-            return
-
-        # Mouse buttom was pressed on the image. We start panning.
-        self.roi_origin = click_position
-        self.__zoomSelectionIndicator.setGeometry(QtCore.QRect(self.roi_origin, QtCore.QSize()))
-        self.__zoomSelectionIndicator.show()
+            # Mouse buttom was pressed on the image. We start panning.
+            self.roi_origin = click_position
+            self.__zoomSelectionIndicator.setGeometry(QtCore.QRect(self.roi_origin, QtCore.QSize()))
+            self.__zoomSelectionIndicator.show()
+        elif event.button() == QtCore.Qt.RightButton:
+            # Flag to indicate click occured
+            self.right_button_pressed = True
 
     def mouseMoveEvent(self, event):
         """
@@ -239,12 +290,18 @@ class ImagePlotWidget(RawImageWidget):
         :param event: (QMouseEvent) Parameter holding event details.
         :return: (None)
         """
-        if self.roi_origin is None:
-            return
+        if self.mouse_dialog.mouse_dialog_enabled:
+            self.updateMouseDialog(event)
+
+        # panning zoomed image window
+        if self.is_zoomed() and self.right_button_pressed:
+            self.is_image_panning = True
+            self.pan_zoom_window(event)
 
         # Mouse is moving, while selecting roi. Redraw the selection rectangle.
-        self.__zoomSelectionIndicator.setGeometry(
-            QtCore.QRect(self.roi_origin, event.pos()).normalized())
+        if self.roi_origin is not None:
+            self.__zoomSelectionIndicator.setGeometry(
+                QtCore.QRect(self.roi_origin, event.pos()).normalized())
 
     def mouseReleaseEvent(self, event):
         """
@@ -253,6 +310,13 @@ class ImagePlotWidget(RawImageWidget):
         :param event: (QMouseEvent) Parameter holding event details.
         :return: (None)
         """
+        # Trigger context menu when not panning
+        if event.button() == QtCore.Qt.RightButton:
+            self.right_button_pressed = False
+            if not self.is_image_panning:
+                self.right_button_clicked_signal.emit(event.pos())
+            self.is_image_panning = False
+
         if self.roi_origin is None:
             return
 
@@ -290,6 +354,178 @@ class ImagePlotWidget(RawImageWidget):
 
         # Calculate zoomed image parameters
         self.__calculateZoomParameters(xmin, xmax, ymin, ymax)
+    
+    def wheelEvent(self, event):
+        """
+        This method checks the angleDelta to determine the direction of the scroll to then
+        set zoom region
+
+        :param event: (QMouseEvent) Parameter holding event details.
+        :return: (None)
+        """
+        # Calculate the position of the mouse cursor relative to the widget
+        mouse_x = event.x()
+        mouse_y = event.y()
+
+        # Get current zoom parameters
+        xOffset, yOffset, width, height = self.get_zoom_region()
+
+        # Calculate the scale factor for zooming based on the mouse wheel delta
+        scale_factor = 0.95 if event.angleDelta().y() > 0 else 1/0.95
+
+        # Calculate the new width and height after zooming
+        new_width = width * scale_factor
+        new_height = height * scale_factor
+
+        # Limit the minimum and maximum size after zooming
+        min_size = 4
+        
+        # Zooming out/in
+        if (scale_factor > 1):
+            max_width_size = self.x
+            max_height_size = self.y
+            new_width = math.ceil(max(min(new_width, max_width_size), min_size))
+            new_height = math.ceil(max(min(new_height, max_height_size), min_size))
+        else:
+            max_width_size = width
+            max_height_size = height
+            new_width = int(max(min(new_width, max_width_size), min_size))
+            new_height = int(max(min(new_height, max_height_size), min_size))
+
+        # Calculate the relative position of the mouse cursor within the widget
+        relative_mouse_x = mouse_x / width
+        relative_mouse_y = mouse_y / height
+
+        # Calculate the new position of the mouse cursor after zooming
+        new_mouse_x = relative_mouse_x * new_width
+        new_mouse_y = relative_mouse_y * new_height
+
+        # Calculate the new position of the widget after zooming to keep the mouse cursor centered
+        new_x = int(xOffset + mouse_x - new_mouse_x)
+        new_y = int(yOffset + mouse_y - new_mouse_y)
+        if new_x < 0:
+            new_x = 0
+        elif new_x > self.x:
+            new_x = self.x
+        if new_y < 0:
+            new_y = 0
+        elif new_y > self.y:
+            new_y = self.y
+
+        # Set the new width, height, and position of the widget
+        self.set_zoom_region(new_x, new_y, new_width, new_height)
+
+    def updateMouseDialog(self, event):
+        """
+        This method continuously displays and updates the dialog box beneath the mouse cursor.
+
+        :param event: (QMouseEvent) Parameter holding event details.
+        :return: (None)
+        """
+        # Return if there is no image data
+        if not self.x or not self.y:
+            return
+        
+        # Flag to indicate mouse dialog has been launched
+        self.mouse_dialog.mouse_dialog_launched = True
+        
+        # Get the current image size on screen
+        img_width, img_height, pixel_size = self.calc_img_size_on_screen()
+
+        # Get the current mouse position on the widget
+        mouse_position = event.pos()
+
+        # Get zoom parameters
+        xOffset, yOffset, width, height = self.get_zoom_region()
+
+        # Calculate pixel coordinates
+        self.mouse_dialog.pix_x = int(mouse_position.x()//pixel_size) + xOffset
+        self.mouse_dialog.pix_y = int(mouse_position.y()//pixel_size) + yOffset
+        if self.mouse_dialog.pix_x >= xOffset + width:
+            self.mouse_dialog.pix_x = xOffset + width - 1
+        if self.mouse_dialog.pix_y >= yOffset + height:
+            self.mouse_dialog.pix_y = yOffset + height - 1
+
+        # Update the textbox location inside the widget screen as the mouse moves
+        self.mouse_dialog.x_coor = mouse_position.x()+5
+        self.mouse_dialog.y_coor = mouse_position.y()+10
+        
+        # Calculate the intensity coordinates corrected for zoom
+        self.mouse_dialog.int_x = self.mouse_dialog.pix_x - xOffset
+        self.mouse_dialog.int_y = self.mouse_dialog.pix_y - yOffset
+
+        # if mouse cursor outside image window, hide mouse dialog
+        if (mouse_position.y() >= (height * pixel_size)) or (mouse_position.x() >= (width * pixel_size)):
+            self.mouse_dialog.mouse_dialog_hide = True
+        else:
+            self.mouse_dialog.mouse_dialog_hide = False
+
+        # Set the dialog box paramters
+        self.setup_mouse_textbox()
+
+        # Flip the textbox location when approaching image widget screen boundary
+        if (mouse_position.x() >= (width * pixel_size - self.mouse_dialog.max_textbox_width)):
+            self.mouse_dialog.x_coor = self.mouse_dialog.x_coor - self.mouse_dialog.max_textbox_width - 20
+        if (mouse_position.y() >= (height * pixel_size - 50)):
+            self.mouse_dialog.y_coor = self.mouse_dialog.y_coor - 55
+        
+        # Set layout
+        self.mouse_dialog.layout.setContentsMargins(self.mouse_dialog.x_coor, self.mouse_dialog.y_coor,int(img_width), int(img_height))
+        self.mouse_dialog.layout.addWidget(self.mouse_dialog.textbox)
+
+    def setup_mouse_textbox(self):
+        """
+        This method set the mouse textbox size and text within based on the image color and data type.
+
+        """
+        # if mouse cursor outside image window, hide mouse dialog
+        if self.mouse_dialog.mouse_dialog_hide or self.underMouse() is False:
+            self.mouse_dialog.textbox.setHidden(True)
+            return
+        else:
+            self.mouse_dialog.textbox.setHidden(False)
+        
+        if self.color_mode == COLOR_MODE_MONO:
+            # Get the monochromatic intensity value at the particular pixel and display in the correct format
+            try:
+                intensity = self.last_displayed_image.image[self.mouse_dialog.int_x][self.mouse_dialog.int_y]
+                if intensity >= 999999:
+                    intensity = "{:e}".format(intensity)
+                self.mouse_dialog.textbox.setText(f"({self.mouse_dialog.pix_x}, {self.mouse_dialog.pix_y})\nI: {intensity}")
+                # Set textbox size
+                if (self.mouse_dialog.textbox.fontMetrics().boundingRect(self.mouse_dialog.textbox.toPlainText()).width() > self.mouse_dialog.max_textbox_width):
+                    self.mouse_dialog.max_textbox_width = self.mouse_dialog.textbox.fontMetrics().boundingRect(self.mouse_dialog.textbox.toPlainText()).width()
+                self.mouse_dialog.textbox.setFixedSize(self.mouse_dialog.max_textbox_width+5, 45)
+            except Exception as e:
+                logging.getLogger().error('Error displaying mouse dialog: %s', str(e))
+        else:
+            # Get the RGB intensity values at the particular pixel and display in the correct format
+            try:
+                r_intensity = self.last_displayed_image.image[self.mouse_dialog.int_x][self.mouse_dialog.int_y][0]
+                if r_intensity >= 999999:
+                    r_intensity = "{:e}".format(r_intensity)
+                g_intensity = self.last_displayed_image.image[self.mouse_dialog.int_x][self.mouse_dialog.int_y][1]
+                if g_intensity >= 999999:
+                    g_intensity = "{:e}".format(g_intensity)
+                b_intensity = self.last_displayed_image.image[self.mouse_dialog.int_x][self.mouse_dialog.int_y][2]
+                if b_intensity >= 999999:
+                    b_intensity = "{:e}".format(b_intensity)
+                
+                # Set textbook text and size
+                if self._inputType == 'ubyteValue' or self._inputType == 'byteValue' or self._inputType == 'ushortValue' or self._inputType == 'shortValue':
+                    self.mouse_dialog.textbox.setText(f"({self.mouse_dialog.pix_x}, {self.mouse_dialog.pix_y})\nR: {r_intensity}, G: {g_intensity}, B: {b_intensity}")
+                    if (self.mouse_dialog.textbox.fontMetrics().boundingRect(self.mouse_dialog.textbox.toPlainText()).width() > self.mouse_dialog.max_textbox_width):
+                        self.mouse_dialog.max_textbox_width = self.mouse_dialog.textbox.fontMetrics().boundingRect(self.mouse_dialog.textbox.toPlainText()).width()
+                    self.mouse_dialog.textbox.setFixedSize(self.mouse_dialog.max_textbox_width+5, 45)
+                else:
+                    self.mouse_dialog.textbox.setText(f"({self.mouse_dialog.pix_x}, {self.mouse_dialog.pix_y})\nR: {r_intensity}")
+                    if (self.mouse_dialog.textbox.fontMetrics().boundingRect(self.mouse_dialog.textbox.toPlainText()).width() > self.mouse_dialog.max_textbox_width):
+                        self.mouse_dialog.max_textbox_width = self.mouse_dialog.textbox.fontMetrics().boundingRect(self.mouse_dialog.textbox.toPlainText()).width()
+                    self.mouse_dialog.textbox.setText(f"({self.mouse_dialog.pix_x}, {self.mouse_dialog.pix_y})\nR: {r_intensity},\nG: {g_intensity},\nB: {b_intensity}")
+                    self.mouse_dialog.textbox.setFixedSize(self.mouse_dialog.max_textbox_width+5, 85)
+            except Exception as e:
+                logging.getLogger().error('Error displaying mouse dialog: %s', str(e))
+
 
     def set_display_queue_size(self, new_size):
         """
@@ -405,6 +641,38 @@ class ImagePlotWidget(RawImageWidget):
                 self.display(self.data, zoomUpdate=True)
             except Exception as e:
                 logging.getLogger().error('Error displaying data: %s', str(e))
+    
+    def pan_zoom_window(self, event):
+        """
+        This method allows a user to move the zoom window around the image by dragging 
+        their mouse while holding down the right button
+
+        :param event: (QMouseEvent) Parameter holding event details.
+        :return: (None)
+        """
+        # Get the current image size on screen
+        _, _, pixel_size = self.calc_img_size_on_screen()
+
+        # Get the current mouse position on the widget
+        new_mouse_position = event.pos()
+
+        # Get zoom parameters
+        xOffset, yOffset, width, height = self.get_zoom_region()
+
+        # Calculate new offset coordinates for zoom window
+        new_x = int(new_mouse_position.x()//pixel_size) - int(self.click_x//pixel_size) + xOffset
+        new_y = int(new_mouse_position.y()//pixel_size) - int(self.click_y//pixel_size) + yOffset
+        if new_x < 0:
+            new_x = 0
+        elif new_x > self.x - width:
+            new_x = self.x - width
+        if new_y < 0:
+            new_y = 0
+        elif new_y > self.y - height:
+            new_y = self.y - height
+
+        # Set the new width, height, and position of the widget
+        self.set_zoom_region(new_x, new_y, width, height)
 
     def set_freeze(self, flag):
         """
@@ -484,7 +752,7 @@ class ImagePlotWidget(RawImageWidget):
         image_width = pixel_size * width
         image_height = pixel_size * height
 
-        return image_width, image_height
+        return image_width, image_height, pixel_size
 
     def wait(self):
         """
@@ -968,7 +1236,16 @@ class ImagePlotWidget(RawImageWidget):
         levels = [image.black, image.white]
         self.setImage(image.image, levels = levels)
         if self.image_profile_widget is not None:
-            self.image_profile_widget.plot(*self.calc_img_size_on_screen())
+            img_width, img_height, _ = self.calc_img_size_on_screen()
+            self.image_profile_widget.plot(img_width, img_height)
         if image.new:
             self.frames_displayed += 1
         self.last_displayed_image = image
+
+        # If the mouse dialog has been launched, update the intensity value with every image update
+        if (self.mouse_dialog.mouse_dialog_launched):
+            # Set the dialog box paramters
+            self.setup_mouse_textbox()
+
+            # Set layout
+            self.mouse_dialog.layout.addWidget(self.mouse_dialog.textbox)